modbus_client/requirements.txt
paho-mqtt==2.1.0
pymodbus==3.6.9
